requirements.txt
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
coverage==7.1.0
dataclasses-json==0.5.7
exceptiongroup==1.1.0
Flask==2.2.2
flask-swagger-ui==4.11.1
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
packaging==23.0
pluggy==1.0.0
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytest==7.2.1
python-dotenv==0.21.1
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
tomli==2.0.1
typing-extensions==4.4.0
typing-inspect==0.8.0
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.12.0
